adding protected route

+++ src/components/ApplicationIndex/ApplicationIndex.jsx
@@ -2,6 +2,7 @@ import { useEffect, useState } from 'react'
 import axios from 'axios'
 import { authRequest, getUserFromToken, clearTokens } from "../../lib/auth"
 import './ApplicationIndex.css'
+import Swal from 'sweetalert2'
 
 function ApplicationIndex() {
 
@@ -24,10 +25,28 @@ function ApplicationIndex() {
 
     async function handleApplicationDelete(applicationId){
         try {
-            const response = await authRequest({method: 'delete', url:`http://127.0.0.1:8000/api/applications/${applicationId}/`})
-            setApplications((prevState) => {
-                return prevState.filter((application) => application.id !== applicationId)
-            })
+            Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+            }).then((result) => {
+            if (result.isConfirmed) {
+                const response = authRequest({method: 'delete', url:`http://127.0.0.1:8000/api/applications/${applicationId}/`})
+                setApplications((prevState) => {
+                    return prevState.filter((application) => application.id !== applicationId)
+                })
+                Swal.fire({
+                title: "Deleted!",
+                text: "Your file has been deleted.",
+                icon: "success"
+                });
+            }
+            });
+            
         } catch (error) {
             setErrors(error.response.data.error)
         }
